feat: modularized application structure
