+++ src/token_validation.py
@@ -1,0 +1,14 @@
+def validate_token_burn(consumer_address, token_amount):
+    """
+    Validate token burn for access.
+
+    Parameters:
+    - consumer_address (str): Address of the consumer performing token burn.
+    - token_amount (int): Amount of tokens burned.
+
+    Returns:
+    - bool: True if token burn is valid, False otherwise (placeholder for actual token validation).
+    """
+    # Don't implement this function.
+    # It is only used for simulation purposes.
+    return True  # Simulate valid token burn
